requirements.txt
altgraph==0.17.2
anyio==3.6.2
APScheduler==3.6.3
schedule==1.1.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.0
future==0.18.2
google-auth==1.12.0
google-auth-oauthlib==0.5.3
gspread==5.7.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
macholib==1.15.2
oauthlib==3.2.2
pip==21.2.4
platformdirs==2.5.4
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pyrogram==2.0.61
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.0
setuptools==58.0.4
six==1.15.0
sniffio==1.3.0
TgCrypto==1.2.5
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.16.7
wheel==0.37.0
